webkit/data/layout_tests/chrome/fast/dom/domListEnumeration-expected.txt

This tests enumerating the elements of DOM lists.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



[object NodeList]
PASS resultArray.length is 5
PASS resultArray[0].i is '0'
PASS resultArray[0].item is nodeList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is nodeList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is nodeList.item(2)
PASS resultArray[3].i is 'item'
PASS resultArray[3].item is nodeList.item
PASS resultArray[4].i is 'length'
PASS resultArray[4].item is nodeList.length

[object HTMLCollection]
PASS resultArray.length is 7
PASS resultArray[0].i is '0'
PASS resultArray[0].item is htmlCollection.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is htmlCollection.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is htmlCollection.item(2)
PASS resultArray[3].i is 'item'
PASS resultArray[3].item is htmlCollection.item
PASS resultArray[4].i is 'length'
PASS resultArray[4].item is htmlCollection.length
PASS resultArray[5].i is 'namedItem'
PASS resultArray[5].item is htmlCollection.namedItem
PASS resultArray[6].i is 'tags'
PASS resultArray[6].item is htmlCollection.tags

[object NamedNodeMap]
PASS resultArray.length is 11
PASS resultArray[0].i is '0'
PASS resultArray[0].item is namedNodeMap.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is namedNodeMap.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is namedNodeMap.item(2)
PASS resultArray[3].i is 'getNamedItem'
PASS resultArray[3].item is namedNodeMap.getNamedItem
PASS resultArray[4].i is 'getNamedItemNS'
PASS resultArray[4].item is namedNodeMap.getNamedItemNS
PASS resultArray[5].i is 'item'
PASS resultArray[5].item is namedNodeMap.item
PASS resultArray[6].i is 'length'
PASS resultArray[6].item is namedNodeMap.length
PASS resultArray[7].i is 'removeNamedItem'
PASS resultArray[7].item is namedNodeMap.removeNamedItem
PASS resultArray[8].i is 'removeNamedItemNS'
PASS resultArray[8].item is namedNodeMap.removeNamedItemNS
PASS resultArray[9].i is 'setNamedItem'
PASS resultArray[9].item is namedNodeMap.setNamedItem
PASS resultArray[10].i is 'setNamedItemNS'
PASS resultArray[10].item is namedNodeMap.setNamedItemNS

[object HTMLFormElement]
PASS resultArray.length is 114
PASS resultArray[0].i is '0'
PASS resultArray[0].item is document.getElementsByTagName('select')[0]
PASS resultArray[1].i is '1'
PASS resultArray[1].item is document.getElementsByTagName('select')[1]
PASS resultArray[2].i is '2'
PASS resultArray[2].item is document.getElementsByTagName('select')[2]
PASS resultArray[3].i is 'ATTRIBUTE_NODE'
PASS resultArray[3].item is htmlFormElement.ATTRIBUTE_NODE
...ellided the remaining 109

[object HTMLSelectElement]
PASS resultArray.length is 117
PASS resultArray[0].i is '0'
PASS resultArray[0].item is document.getElementsByTagName('option')[0]
PASS resultArray[1].i is '1'
PASS resultArray[1].item is document.getElementsByTagName('option')[1]
PASS resultArray[2].i is '2'
PASS resultArray[2].item is document.getElementsByTagName('option')[2]
PASS resultArray[3].i is 'ATTRIBUTE_NODE'
PASS resultArray[3].item is htmlSelectElement.ATTRIBUTE_NODE
...ellided the remaining 113

[object StyleSheetList]
PASS resultArray.length is 5
PASS resultArray[0].i is '0'
PASS resultArray[0].item is styleSheetList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is styleSheetList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is styleSheetList.item(2)
PASS resultArray[3].i is 'item'
PASS resultArray[3].item is styleSheetList.item
PASS resultArray[4].i is 'length'
PASS resultArray[4].item is styleSheetList.length

[object CSSRuleList]
PASS resultArray.length is 5
PASS resultArray[0].i is '0'
PASS resultArray[0].item is cssRuleList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is cssRuleList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is cssRuleList.item(2)
PASS resultArray[3].i is 'item'
PASS resultArray[3].item is cssRuleList.item
PASS resultArray[4].i is 'length'
PASS resultArray[4].item is cssRuleList.length

[object CSSStyleDeclaration]
PASS resultArray.length is 14
PASS resultArray[0].i is '0'
PASS resultArray[0].item is cssStyleDeclaration.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is cssStyleDeclaration.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is cssStyleDeclaration.item(2)
PASS resultArray[3].i is 'cssText'
PASS resultArray[3].item is cssStyleDeclaration.cssText
PASS resultArray[4].i is 'getPropertyCSSValue'
PASS resultArray[4].item is cssStyleDeclaration.getPropertyCSSValue
PASS resultArray[5].i is 'getPropertyPriority'
PASS resultArray[5].item is cssStyleDeclaration.getPropertyPriority
PASS resultArray[6].i is 'getPropertyShorthand'
PASS resultArray[6].item is cssStyleDeclaration.getPropertyShorthand
PASS resultArray[7].i is 'getPropertyValue'
PASS resultArray[7].item is cssStyleDeclaration.getPropertyValue
PASS resultArray[8].i is 'isPropertyImplicit'
PASS resultArray[8].item is cssStyleDeclaration.isPropertyImplicit
PASS resultArray[9].i is 'item'
PASS resultArray[9].item is cssStyleDeclaration.item
PASS resultArray[10].i is 'length'
PASS resultArray[10].item is cssStyleDeclaration.length
PASS resultArray[11].i is 'parentRule'
PASS resultArray[11].item is cssStyleDeclaration.parentRule
PASS resultArray[12].i is 'removeProperty'
PASS resultArray[12].item is cssStyleDeclaration.removeProperty
PASS resultArray[13].i is 'setProperty'
PASS resultArray[13].item is cssStyleDeclaration.setProperty

[object CSSValueList]
PASS resultArray.length is 10
PASS resultArray[0].i is '0'
PASS resultArray[0].item is cssValueList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is cssValueList.item(1)
PASS resultArray[2].i is 'CSS_CUSTOM'
PASS resultArray[2].item is cssValueList.CSS_CUSTOM
PASS resultArray[3].i is 'CSS_INHERIT'
PASS resultArray[3].item is cssValueList.CSS_INHERIT
PASS resultArray[4].i is 'CSS_PRIMITIVE_VALUE'
PASS resultArray[4].item is cssValueList.CSS_PRIMITIVE_VALUE
PASS resultArray[5].i is 'CSS_VALUE_LIST'
PASS resultArray[5].item is cssValueList.CSS_VALUE_LIST
PASS resultArray[6].i is 'cssText'
PASS resultArray[6].item is cssValueList.cssText
PASS resultArray[7].i is 'cssValueType'
PASS resultArray[7].item is cssValueList.cssValueType
PASS resultArray[8].i is 'item'
PASS resultArray[8].item is cssValueList.item
PASS resultArray[9].i is 'length'
PASS resultArray[9].item is cssValueList.length

[object MediaList]
PASS resultArray.length is 8
PASS resultArray[0].i is '0'
PASS resultArray[0].item is mediaList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is mediaList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is mediaList.item(2)
PASS resultArray[3].i is 'appendMedium'
PASS resultArray[3].item is mediaList.appendMedium
PASS resultArray[4].i is 'deleteMedium'
PASS resultArray[4].item is mediaList.deleteMedium
PASS resultArray[5].i is 'item'
PASS resultArray[5].item is mediaList.item
PASS resultArray[6].i is 'length'
PASS resultArray[6].item is mediaList.length
PASS resultArray[7].i is 'mediaText'
PASS resultArray[7].item is mediaList.mediaText

PASS successfullyParsed is true

TEST COMPLETE
